refactor(tasks): move task type into feature
